+++ CHANGELOG.md
@@ -6,6 +6,13 @@
 - run failed test support
 - display tests in a better way?
 
+## [0.0.4] - 2018-02-17
+### Added
+- Clear terminal between each run (configure to disable, true by default)
+- Focus terminal on each run (configure to disable, true by default)
+- Add support for --fail-fast in rspec
+- Add command to re-run last failed tests
+
 ## [0.0.3] - 2018-01-04
 ### Added
 - New changelog (woho)